Support both JWT and Opaque token configuration for resource server
Closes gh-19426
